pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "churn-intelligence-system"
version = "3.0.0"
description = "End-to-end Churn Intelligence System with MLflow tracking and ReAct Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Benjamin", email = "benjamin@example.com"},
]

dependencies = [
    # --- Core Data & ML ---
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "imbalanced-learn>=0.11.0",

    # --- Experiment Tracking ---
    "mlflow>=2.12.0",

    # --- Explainability ---
    "shap>=0.44.0",

    # --- Monitoring ---
    "evidently>=0.4.10",

    # --- Visualization ---
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",

    # --- API ---
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",

    # --- Streamlit ---
    "streamlit>=1.31.0",

    # --- Agent (LangGraph + LangChain) ---
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-core>=0.2.0",

    # --- Gradient Boosting (CatBoost) ---
    "catboost>=1.2.0",

    # --- Hyperparameter Optimisation ---
    "optuna>=3.5.0",

    # --- Dimensionality Reduction (optional — gracefully skipped if absent) ---
    "pacmap>=0.7.0",

    # --- Utilities ---
    "python-dotenv>=1.0.0",
    "joblib>=1.3.0",
    "kagglehub>=0.2.0",
    "loguru>=0.7.0",
    "jinja2>=3.1.0",
    "lifelines>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "ipykernel>=6.25.0",
    "nbformat>=5.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "agent*", "api*", "app*", "policy*", "tracking*"]

[tool.black]
line-length = 99
target-version = ["py310"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.isort]
profile = "black"
line_length = 99

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
